requirements.txt
Flask==2.2.5
flask-smorest
Flask-SQLAlchemy
Flask-Migrate
passlib
marshmallow